Puse como integer al id del soporte, al igual que el soporteService, el id el soporte lo puse como integer, ya que como long esta incorrecto. El soporteService esta terminado, pero no se ha probado aun, tal vez haya que revisar y cambiar correctamente algunos a integer o long, segun corresponda para evitar errores

+++ src/main/java/com/ecomarket/cl/ecomarket/controller/SoporteController.java
@@ -58,7 +58,7 @@ public class SoporteController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> eliminar(@PathVariable Long id){
+    public ResponseEntity<?> eliminar(@PathVariable Integer id){
         try {
             soporteService.delete(id);
             return ResponseEntity.noContent().build();
